Run Spring Boot on Tomcat.
