+++ tomcat/docker-compose.yaml
@@ -10,6 +10,16 @@ services:
       - "5432:5432"
     volumes:
       - postgres_data:/var/lib/postgresql/data
+
+  tomcat:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    container_name: todoapp_tomcat
+    ports:
+      - "8080:8080"
+    depends_on:
+      - postgres
       
 volumes:
   postgres_data:
