+++ tomcat/src/main/resources/application.properties
@@ -1,6 +1,8 @@
 spring.application.name=Todo
 
-spring.datasource.url=jdbc:postgresql://localhost:5432/todoapp_db
+spring.jpa.properties.hibernate.transaction.coordinator_class=jdbc
+
+spring.datasource.url=jdbc:postgresql://postgres:5432/todoapp_db
 spring.datasource.username=postgres
 spring.datasource.password=admin
 spring.datasource.driver-class-name=org.postgresql.Driver
@@ -9,6 +11,8 @@ spring.jpa.hibernate.ddl-auto=none
 spring.jpa.show-sql=true
 
 spring.jpa.properties.jakarta.persistence.schema-generation.scripts.action=create-drop
-spring.jpa.properties.jakarta.persistence.schema-generation.scripts.create-target=schema.sql
+spring.jpa.properties.jakarta.persistence.schema-generation.scripts.create-target=scripts/schema.sql
 spring.jpa.properties.jakarta.persistence.schema-generation.scripts.create-source=metadata
-spring.jpa.properties.jakarta.persistence.schema-generation.scripts.drop-target=drop.sql
+spring.jpa.properties.jakarta.persistence.schema-generation.scripts.drop-target=scripts/drop.sql
+
+logging.level.org.hibernate=DEBUG
