+++ tomcat/src/main/resources/scripts/drop.sql
