+++ tomcat/src/main/todo/package.json
@@ -12,6 +12,7 @@
     "react-toastify": "^11.0.2"
   },
   "proxy": "http://localhost:8080",
+  "homepage": "/todo",
   "scripts": {
     "start": "react-app-rewired start",
     "build": "react-app-rewired build",
